Restored old pure JS algorithm
Added multi backends support
