+++ packages/demo/index.html
@@ -17,6 +17,13 @@
         </p>
       </label>
       <div class="parameter">
+        <label for="strokeBackend">Backend</label>
+        <select id="strokeBackend">
+          <option default value="purejs">Pure JS</option>
+          <option value="webgl">WebGL</option>
+        </select>
+      </div>
+      <div class="parameter">
         <label for="strokeColor">Color</label>
         <input type="color" id="strokeColor" value="#ffffff">
       </div>
